Completed Playtime with clock in Gui

+++ CustomSkyblock/src/CustomEssentials/Events/Misc/TimeGenerator.java
@@ -10,20 +10,20 @@ public class TimeGenerator {
 	    
 	    if (s >= 60) {
 	    	mins = Math.floorDiv((int) s, 60);
-	    	s = (float) Math.floor(((s/60) - mins)*60);
+	    	s = (int) Math.floor(((s/60) - mins)*60);
 	    }
 	    if (mins >= 60) {
 	    	hrs = Math.floorDiv((int) s, 60);
-	    	mins = (float) Math.floor(((mins/60) - hrs)*60);
+	    	mins = (int) Math.floor(((mins/60) - hrs)*60);
 	    }
 	    if (hrs >= 60) {
 	    	days = Math.floorDiv((int) s, 60);
-	    	hrs = (float) Math.floor(((hrs/60) - days)*60);
+	    	hrs = (int) Math.floor(((hrs/60) - days)*60);
 	    }
 	    
 		
 		
-	    return (days+"d "+ hrs+"hr "+ mins+"m "+ s+"s ");
+	    return ((int)days+"d "+ (int)hrs+"hr "+ (int)mins+"m "+ (int)s+"s ");
 		
 	}
